requirements.txt
FastAPI
uvicorn
pandas
openpyxl
python-dotenv
requests
twilio